maia-wasm: do not make promise fail if PATCH fails
In the UI patch_* methods, if the PATCH request fails, rather than
making the promise fail instantly, we log as an error the response
text, which contains more information about what went wrong.
Additionally, we consider this a "non-fatal" error for the promise,
so we continue running the promise, skipping work if needed (for
instance, if the PATCH response would normally be used to update
some UI fields) and return Ok from the promise. If other errors happen
during the PATCH or the rest of the promise, we still consider these
fatal errors.

Signed-off-by: Daniel Estévez <daniel@destevez.net>