maia-wasm: avoid PATCHing RX gain in non-manual AGC
When the AGC is not set to manual mode, PATCH requests that try to
set the RX gain will fail. This adds custom onchange method for the
ad9361_rx_gain field that checks if the ad9361_rx_gain_mode is manual,
and only calls the macro-generated onchange method in this case, so
as to avoid having PATCH requests that fail (specially when applying
the saved properties to the server when the app is loaded).

Signed-off-by: Daniel Estévez <daniel@destevez.net>