maia-wasm: refactor to improve waterfall reusability
Inspired by an experiment by bastibl about using the maia-wasm
waterfall in FutureSDR, this improves the reusability of the
waterfall by reorganizing the code of the start function by
maia-wasm into separate functions, so that these can be called
in other use cases. An example of a standalone waterfall will
show how to use these effectively.

The canvas resizing functions have been moved from Ui to
WaterfallInteraction in order to support use cases in which Ui is
not present, because the forms and REST API are missing. Moreover,
now WaterfallInteraction sets up its own callbacks on construction,
in the same way as Ui.

The crate console_error_panic_hook has been added to aid in debugging.

The wasm init function only registers the panic hook. Afterwards it
is necessary to call maia_wasm_start() or another function to initialize
the desired application.

Signed-off-by: Daniel Estévez <daniel@destevez.net>