maia-wasm: add standalone waterfall example
This adds a crate that contains an example showing how to use the
maia-wasm waterfall in a standalone way. It is inspired by some
experiments done by bastibl about using the waterfall with FutureSDR.

The example is also intended to be shown as a demo in the Maia SDR
webpage.

Signed-off-by: Daniel Estévez <daniel@destevez.net>