maia-wasm: fix wrong center in waterfall_interaction
All the handling of pointer events in waterfall_interaction is done
using client coordinates. However, it is assumed that the origin of
the waterfall canvas in client coordinates is zero. This is not
necessarily the case, so the origin needs to be subtracted.

This bug caused a wrong center to be used for mouse wheel and pinch
gestures when the canvas left edge was far from the left edge of the
client (typically the viewport left edge).

Signed-off-by: Daniel Estévez <daniel@destevez.net>