maia-wasm: let waterfall grow to occupy empty viewport space
This removes the blank space that can appear below the UI form when
the viewport is wide, by letting the waterfall grow to occupy the
remaining space.

A flexbox containing the waterfall canvas and the UI form is added.
This flexbox is used to control the waterfall size, which has a minimum
of 75vh (this has been reduced from 80vh) but can grow to fill the
remainig space of the flexbox minimum height, which is 100vh.

The UI form can cause the flexbox to overflow if the screen is narrow
enough that the form doesn't fit in 25vh.

Signed-off-by: Daniel Estévez <daniel@destevez.net>