maia-wasm: fix frequency change by waterfall scroll
60db8ffcfd704e37346911bca543e7a22eb34c32 broke the RX LO frequency change
by scrolling the waterfall. The reason for this is that by the time that
the onpointermove callback was created in WaterfallInteraction, self.ui
had not yet been set, so the callback would always see this as None, even
if we would update it later on.

The type of WaterfallInteraction::ui is now changed to an Rc<RefCell<>>,
so that the callback "sees" the change that happens later on, when
WaterfallInteraction::set_ui is called.

Signed-off-by: Daniel Estévez <daniel@destevez.net>