maia-httpd,maia-wasm: prepend timestamp to recording filename
This adds an option to prepend the date and time to the IQ recording
filename, as suggested in #3.

A prepend_timestamp boolean field in the REST API controls whether
prepending mode is active. In prepending mode, when a recording is
started, the current filename is examined. If the beginning of the
filename is formatted as a timestamp, this timestamp is stripped.
In other case, the filename is left as is. Next, the timestamp
corresponding to the recording start (the same which is used for the
SigMF metadata) is formatted as %Y-%m-%d-%H-%M-%S_ and prepended
to the filename.

Signed-off-by: Daniel Estévez <daniel@destevez.net>