maia-http,maia-wasm: add recording maximum duration
This adds a maximum duration field to the recorder REST API, which
specifies the maximum duration for the next recording in seconds.
A maximum duration of 0 indicates unlimited maximum duration
(only limited by the RAM buffer size).

If the maximum duration is set to a value t > 0, the recording will
be automatically stopped by maia-httpd after t + 0.1 seconds. The
extra 0.1 seconds is added just in case the CPU timer is faster than
the ADC sampling clock, or not very accurate, so that there is always
the guarantee that at least the required number of samples have been
recorded to RAM.

When the SigMF file is assembled for download, the SigMF file will
be formed with exactly the number of samples corresponding to the
maximum duration (this feature can even be used after the recording
has been done, to "cut short" a longer recording).

Signed-off-by: Daniel Estévez <daniel@destevez.net>