maia-hdl: add Register changed access modes
The RWchanged and Wchanged access modes are like RW and W, but the
include an additional "changed" output that is pulsed for one clock
cycle when the register is written. The changed output can be
access with the changed() method of Registers.

These register access are required by some control ports of the
fosphor IP core.

Signed-off-by: Daniel Estévez <daniel@destevez.net>