maia-httpd: fix panic with uClibc and rustc >= 1.69.0
This includes a workaround for
https://github.com/rust-lang/rust/issues/112488
which causes a panic when accessing command line arguments in
rustc >= 1.69.0. Since accessing command line arguments is anyhow
broken in uClibc, we just use default arguments when building with
the uclibc feature (which is a default feature to simplify building
for the intended target, which is the Pluto firmware).

Signed-off-by: Daniel Estévez <daniel@destevez.net>