maia-hdl: add project for Pluto+
This adds a project for the Pluto+. The differences between the Pluto+
and the ADI ADALM Pluto Vivado projects are collected in this diff:

https://github.com/plutoplus/plutoplus/blob/master/patches/hdl.diff

The .xdc is completely different because the FPGA uses a different
package and pinout. There are also a few differences in the Processing
System settings, because the Pluto+ includes Ethernet and SD card,
which the ADALM Pluto doesn't have.

To avoid code duplication, the system_bd.tcl of the pluto project is
re-used. Some if statements that depend on the existence of a
'plutoplus' variable are used to apply the settings specific to the
Pluto+.

Signed-off-by: Daniel Estévez <daniel@destevez.net>