maia-wasm: fix large number of frequency ticks and labels
The number of frequencies that was generated for the ticks and
labels in the waterfall was too large (above 2^16). Since we use
a u16 element array buffer, we were not drawing most of the generated
frequencies due to overflows when converting to u16.

This bug was noticed because at 45 Msps the ticks wouldn't draw at
all.

This decreases the depth of step levels to a more reasonable value,
which seems to guarantee that we don't generate more than 2^16
vertices for the ticks nor for the labels. Asserts have been added
to check that this is indeed the case.

The sample rate at which we get more frequencies seems to be 20.7 Msps,
where we get 25555 frequencies. At 20.8 Msps we get 9111 frequencies.
These two are perhaps the extreme cases of this frequency gridding
algorithm.

This fixes #73.

Signed-off-by: Daniel Estévez <daniel@destevez.net>