maia-wasm: cull waterfall frequencies outside passband
This eliminates from the list of frequency ticks the frequencies
which fall outside the passband, reducing the number of elements
in the VAOs used for the frequency axis. This change was marked
as a TODO, because it is a good optimization. However, it has
been implemented specifically to address #79.

In that issue we have found that for large horizontal resolutions
(above 2800 px approx.), the waterfall would panic because the
number of generated frequencies is too large (the VAOs are indexed
by uint16, so this sets a limitation). This change improves the
situation, allowing horizontal resolutions of up to approx 5600 px
(the exact value depends on the sample rate).

Larger resolutions would still crash, but solving this would require
a different strategy. In addition, with very large resolutions maybe
it is not reasonable to use the same size for the frequency labels
(it is specified in pixels of text height). If the size of the text
in pixels is proportional to the resolution, then there would be no
problem. The only reason why there is a problem is that now the text
size in px is fixed, so larger horizontal resolutions are able to
squeeze together more frequency labels for a given zoom factor.

Signed-off-by: Daniel Estévez <daniel@destevez.net>