maia-wasm: fix clippy lints
Some clippy lints have appeared in the latest version of the Rust
toolchain. These lints are above an unncessary
 let patch = patch;

This was used before because it is necessary to move patch into
the async block, but the block only depends on &patch, so the
compiler didn't seem smart enough to figure out the need to make
the move on its own. Apparently now it does and the let is redundant.

Signed-off-by: Daniel Estévez <daniel@destevez.net>