maia-wasm: limit depth of frequency labels to avoid panic
Limit the depth of the generated list of frequency ticks and labels
to avoid having a list that is too long (which happens on very large
resolutions).

This complements the culling of frequencies outside the passband
introduced in the previous commit by limiting the length of the
frequency list to avoid panic in all cases. Rather than cutting the
frequency list at the maximum length, it is cut at a constant depth.

Signed-off-by: Daniel Estévez <daniel@destevez.net>