maia-httpd,maia-wasm: add IQEngine viewer
This adds support for using IQEngine to view the IQ recording stored
in RAM. The functionality is implemented as follows:

- The iqengine index.html file must be installed in iqengine/index.html.
It is served as a static file with the URL
/view/api/maiasdr/maiasdr/recording. This URL is compliant with the
scheme used by the IQEngine client for API datasets.
- The assets for the IQEngine client must be installed in iqengine/ and
compressed in LZ4. These are served by a performing decompression on the
fly. Compression is used to save ramfs space, since otherwise Linux runs
out of memory while booting.
- GET methods for the meta, iq-data and minimap-data IQEngine backend
API have been implemented to obtain data from the recording in RAM.
The IQEngine client seamlessly accesses these API methods as if it was
talking to the IQEngine backend.

While doing this, part of the IQ recording httpd code has been
refactored, as it is also used to implement the IQEngine backend API.

Additionally, a button with a link to
/view/api/maiasdr/maiasdr/recording is added in the maia-wasm recording
dialog.

A current limitation of this approach is that the recording is mmap()'ed
for each IQEngine API request, which causes the maia-kmod to invalidate
the cache each time.

Signed-off-by: Daniel Estévez <daniel@destevez.net>