maia-httpd: refactor recording mmap
This refactors how mmap'ing of the IQ recording buffer is done.
Instead of performing an ad-hoc mmap() for each request to serve,
the recording is always mmap()'ed, and requests use this unique
mapping. A RwLock is used to prevent starting a recording while
the current recording is being read, and reading the recording
while the recording is in progress.

When a recording finishes, a new mmap() is done in order to
invalidate the cache, and the old mmap() is unmapped. This approach
is preferrable to modifying the kernel module to include an ioctl
that performs cache invalidation, because the kernel module allows
the buffer to be mmap'ed several times, and each mapping may only
map a small part of the buffer. Since cache invalidation needs to
use virtual addresses for the L1 cache, it is not clear how to
implement cache invalidation with an ioctl() without sacrificing
the flexibility of allowing multiple small mappings.

Signed-off-by: Daniel Estévez <daniel@destevez.net>