maia-httpd,maia-wasm: add version information page
This adds a page /version that lists version information about
maia-httpd, maia-hdl and maia-wasm. maia-httpd has the information
for itself filled in at compile time, and reads /etc/libiio.ini
to get the firmware version, and the maia-hdl IP core registers
to get the maia-hdl version. It generates a page with this information
and some <span>'s that are filled in by JavaScript that calls
maia-wasm functions that return the maia-wasm version information.

Signed-off-by: Daniel Estévez <daniel@destevez.net>