add DDC decimator
This adds the decimator for a DDC to maia-hdl, and the changes to
the other maia-sdr components required to support it. The mixer/NCO
of the DDC is still missing, and it will be implemented later.

The decimator is formed by a cascade of 3 polyphase FIRs. The design
is inspired by the paper "Optimum FIR Digital Filter Implementation
for Decimation, Interpolation, and Narrow-Band Filtering", by Crochiere
and Rabiner.

The decimator has an input of 12 bits (matching the AD9361 ADC)
and output of 16 bits. This change has required adapting the
spectrometer and IQ recorder to 16 bit input.

The FFT now uses a new complex multiplier that can work when one
of the inputs is wider than 18 bits. The SpectrumIntegrator now
uses a floating-point representation, since the 22-bit output of
the FFT has too much dynamic range to use unscaled representation.
The values and exponents are written to the DDR ring buffer, so
maia-httpd needs to be aware of this representation when converting
to f32.

The IQ recorder can now record in 16 bit, 12 bit and 8 bit mode.
When recording from the AD9361 directly in 16 bit mode, the 4 LSBs
are set to zero.

The REST API in maia-httpd used to configure the DDC is low-level at
the moment. The coefficients of each FIR filter are specified in a
PUT request. A higher-level API that specifies the total decimation
and some design requirements (stopband attenuation, etc.) and uses
Remez to solve for optimal coefficients and figures out optimal
decimation distribution over the 3 FIRs is to be added.

Signed-off-by: Daniel Estévez <daniel@destevez.net>