add DDC
This adds a DDC to maia-hdl, and the changes to the other maia-sdr
components required to support it. The DDC is composed of a mixer
and a decimator. The mixer uses a complex multiplier with 1 DSP
running with a 3x clock, and a BRAM to store the complex exponential.

The decimator is formed by a cascade of 3 polyphase FIRs. The design
is inspired by the paper "Optimum FIR Digital Filter Implementation
for Decimation, Interpolation, and Narrow-Band Filtering", by Crochiere
and Rabiner. The decimator runs with a 3x clock and uses 10 DSPs.

The decimator has an input of 12 bits (matching the AD9361 ADC)
and output of 16 bits. This change has required adapting the
spectrometer and IQ recorder to 16 bit input.

The FFT now uses a new complex multiplier that can work when one
of the inputs is wider than 18 bits. The SpectrumIntegrator now
uses a floating-point representation, since the 22-bit output of
the FFT has too much dynamic range to use unscaled representation.
The values and exponents are written to the DDR ring buffer, so
maia-httpd needs to be aware of this representation when converting
to f32.

The IQ recorder can now record in 16 bit, 12 bit and 8 bit mode.
When recording from the AD9361 directly in 16 bit mode, the 4 LSBs
are set to zero.

In maia-httpd, the Parks-McClellan algorithm, implemented in the
pm-remez crate, is used to design the FIR filters for the decimator.
The REST API has methods to submit design requirements, which get
solved with Parks-McClellan if possible, or to directly specify the
FIR filter taps for custom uses.

Building pm-remez requires a custom cross Docker container to build
OpenBLAS, and a patch to cross-compile ndarray-linalg for armv7.

Additionally, in maia-httpd, the application state has been refactored.
Now all the parts of the code that need access part of the state get
an AppState struct that contains the whole state, instead of having
ad-hoc structures that contain only the part of the state that is needed.
This simplifies the code and makes AppState very cheap to clone
in Axum handlers, because it is a single Arc, instead of a struct
containing multiple Arc's.

When making a large change in the spectrum integrator input sample
rate, which happens for instance when enabling or disabling the DDC
with a large decimation factor, the current integration can take
a long time to finish. To avoid this, an abort function has been
added that forces the current integration to terminate.

In maia-wasm, if the power measured by the spectrum integrator
in a particular bin was zero, this was shown with an ugly yellowish
color in the waterfall, because the log10 of the power would give
-infty. Now this is avoided by assigning 0 dB of power in this
case. The dB scale is adjusted so that non-zero values will all give
positive dB values, so this choice of 0 dB to represent zero power
is acceptable. It would saturate as the waterfall min value.

The Vivado implementation strategy has been changed to improve
timing results.

Since the DDC UI requires many input fields, the recording dialog
has been converted into a settings dialog that uses tabs to configure
each functionality. The UI elements that are most frequently used
have remained below the waterfall, but the rest have been moved to
the settings dialog.

The JSON format for errors has now been defined in maia-json. There
is a suggested action for each error. maia-wasm uses this suggested
action to either log the error to the console, show it in a dialog
that is similar to the alert() JavaScript function, or ignore it.

Signed-off-by: Daniel Estévez <daniel@destevez.net>