maia-httpd,maia-wasm: set default decimation to 20
Previously, the DDC was initialized to a "dummy" configuration with
only the first FIR enabled with a decimation of 2 and a single
zero tap. This gave problems with the web UI because if there were
no preferences indicating a higher decimation factor, then it would
request a DDC design for a decimation factor of 2, which would fail
and show an error message.

Instead of initializing the DDC to a dummy configuration, this
initializes the DDC to a valid design for a decimation of 20. This
is relatively fast to compute (about 0.5 seconds) and can be used
for sample rates above 61.44 Msps. The web UI also has the decimation
factor set to 20 as default in the preferences.

Signed-off-by: Daniel Estévez <daniel@destevez.net>