maia-wasm: add setting to toggle showing DDC in waterfall
This adds a setting to the waterfall settings tab that toggles
whether the DDC passband is shown as a shaded bar in the waterfall
or not. Disabling showing the DDC passband can be useful when there
is no intention to use the DDC.

This also fixes a bug in the waterfall interaction. Before checking
which object is being dragged (the DDC passband or the waterfall),
it is checked whether the DDC passband is being shown, since otherwise
it doesn't make sense to drag it.

Signed-off-by: Daniel Estévez <daniel@destevez.net>