improve user feedback when stopping recording
Since the recorder takes a fraction of a second to stop after it
has been commanded to, this adds a new recorder state "stopping"
in maia-json. In maia-wasm, when the recorder is stopping, the
button shows the text "Stopping", it is colored in yellow and it
is not clickable. This gives much better feedback to the user
about what is happening.

Additionally, this swaps the highlight and normal colors of the
recording button in dark mode, so as to have the lighter color
on hover, as is done for the rest of the buttons in dark mode.

Signed-off-by: Daniel Estévez <daniel@destevez.net>