maia-pac: update svd and regenerate code
Regenerate the code with the new SVD (which only changes the version
of the IP core) and a newer version of svd2rust. Run cargo fmt on
the code. Include critical-section as an optional dependency to avoid
lints. The new generated code is API breaking, so there is a major
version bump.

Signed-off-by: Daniel Estévez <daniel@destevez.net>