waterfall-example: add simplified UI
This adds a simplified UI to the waterfall-example. The UI is able
to interact with the waterfall, but not with the maia-httpd REST API,
since this example is standalone and doesn't talk to a maia-httpd.

To do this, a lot of functions, macros and modules in maia_wasm::ui
have been made public. These still need to be documented.

Signed-off-by: Daniel Estévez <daniel@destevez.net>