maia-hdl: add configuration to maia-hdl
This adds a Python-based configuration to maia-hdl. For the time
being, the configuration can only set a few parameters, such as
the addresses for the spectrometer and IQ recorder buffers, as
changing these is needed for maia-iio. In the future this configuration
will also support enabling or disabling optional modules.

The IP core build system has been modified to account for the
configuration. Now a different IP is packaged for each configuration.

Signed-off-by: Daniel Estévez <daniel@destevez.net>