maia-httpd: load certificates from files
Instead of embedding a self-signed SSL certficate, provide command
line parameters to load certificates from files.

A certificate signed with a trusted CA is needed to install maia-wasm
as a PWA (Progressive Web App), because current web browsers require PWAs
to be accessed by SSL with a trusted certificate (not a self-signed
certificate). This opens up a way to install maia-wasm as a PWA, but
it requires the user to trust and add a CA certificate manually in
their device. A startup script will generate a suitable CA and web
server certificates on boot and store them in persistent JFFS flash.
The user can also replace these with their own (by doing scp or
similar).

Signed-off-by: Daniel Estévez <daniel@destevez.net>