maia-httpd: fix regression caused by axum update
In c7d595ef06ea5c368be49a3f18b8669eaca2b4eb there was an API-breaking
update of axum. This built successfully, but was not tested. There are
some runtime errors caused by this update, due to a change in the syntax
for extractors, as described in:

https://github.com/joelparkerhenderson/demo-rust-axum/blob/main/README.md#caution-extractors-syntax-change

This fixes the problem.

Signed-off-by: Daniel Estévez <daniel@destevez.net>